requirements.txt
python-telegram-bot==20.8
beautifulsoup4
requests
apscheduler
